fix: extract gapRpcConfig to utility module and fix isUpdating leak
- Extract getGapRpcConfig from hooks/useGap.ts to utilities/gapRpcConfig.ts
  to fix dependency inversion (utilities should not import from hooks)
- Fix isUpdating never being reset on the unsupported-chain path, which
  permanently blocked future GAP client updates
- Update isCommunityAdmin.ts and projectOwnershipTransfered.ts to import
  from the new utility module
